Alll

+++ Pages/Post-All.aspx.cs
@@ -19,10 +19,10 @@ public partial class Pages_Post_All : BasePage
     protected void Page_Load(object sender, EventArgs e)
     {
         this.setcurenturl();
-        if(!IsPostBack)
+        if (!IsPostBack)
         {
             UserAccounts ac = Session.GetCurrentUser();
-            if(ac==null)
+            if (ac == null)
             {
                 Response.Redirect("http://" + Request.Url.Authority + "/Login.aspx");
             }
@@ -34,14 +34,23 @@ public partial class Pages_Post_All : BasePage
                 }
                 else
                 {
-                    //do somewthing
-                    this.load_dlCategory();
+                    this.AlertPageValid(false, "", alertPageValid, lblPageValid);
+                    if (HasPermission(ac.UserID, FunctionName.PostManager, TypeAudit.View))
+                    {
+                        PostManager.Attributes.Add("class", "row");
+                        //do somewthing
+                        this.load_dlCategory();
+                        this.GetPostPageWise(1, 0);
+                        lblstartindex.Text = ((1 - 1) * PageSize + 1).ToString();
+                        lblendindex.Text = ((((1 - 1) * PageSize + 1) + PageSize) - 1).ToString();
+                    }
+                    else
+                    {
+                        PostManager.Attributes.Add("class", "row display-none");
+                        this.AlertPageValid(true, "Bạn không có quyền xem nội dung này !", alertPageValid, lblPageValid);
+                    }
 
-                    this.GetPostPageWise(1, 0);
-                    lblstartindex.Text = ((1 - 1) * PageSize + 1).ToString();
-                    lblendindex.Text = ((((1 - 1) * PageSize + 1) + PageSize) - 1).ToString();
                 }
-                
             }
         }
     }
@@ -57,9 +66,9 @@ public partial class Pages_Post_All : BasePage
     private void GetPostPageWise(int pageIndex, int categoryID)
     {
         post = new PostBLL();
-        gwPostmanager.DataSource = new ObjectDataSource(); 
+        gwPostmanager.DataSource = new ObjectDataSource();
         int recordCount = 0;
-        if(categoryID<=0)
+        if (categoryID <= 0)
         {
             gwPostmanager.DataSource = post.GetAllPostPageWise(pageIndex, PageSize);
             recordCount = post.CountPost();
@@ -146,7 +155,7 @@ public partial class Pages_Post_All : BasePage
     protected void Page_Changed(object sender, EventArgs e)
     {
         int pageIndex = int.Parse((sender as LinkButton).CommandArgument);
-        this.GetPostPageWise(pageIndex,0);
+        this.GetPostPageWise(pageIndex, 0);
         lblstartindex.Text = ((pageIndex - 1) * PageSize + 1).ToString();
         lblendindex.Text = ((((pageIndex - 1) * PageSize + 1) + PageSize) - 1).ToString();
     }
@@ -160,29 +169,36 @@ public partial class Pages_Post_All : BasePage
                 del.Attributes.Add("onclick", "return confirm('Xóa có thể dẫn tới lỗi hệ thống. Bạn có chắc muốn xóa nữa không ? OK -> Bạn chịu trách nhiệm || Cancel -> coi như không có gì xảy ra !')");
             }
         }
-        catch (Exception)
+        catch (Exception ex)
         {
-
+            Response.Write("<script>alert('" + ex.ToString() + "')</script>");
         }
     }
     protected void gwPostmanager_RowDeleting(object sender, GridViewDeleteEventArgs e)
     {
-        post = new PostBLL();
-        tagsrelationships = new Tags_relationshipsBLL();
-        post_category_relationships = new Post_Category_relationshipsBLL();
-        int postID = Convert.ToInt32((gwPostmanager.Rows[e.RowIndex].FindControl("lblPostID") as Label).Text);
-        bool delTags = this.tagsrelationships.DeleteWithPostId(postID);
-        bool deletepostCT = this.post_category_relationships.DeleteWithPostId(Convert.ToInt32(postID));
-        bool delPost = this.post.DeleteWithPostID(Convert.ToInt32(postID));
-        if (!delTags || !deletepostCT || !delPost)
+        if (HasPermission(Session.GetCurrentUser().UserID, FunctionName.PostManager, TypeAudit.Delete))
         {
-            Response.Write("<script>alert('Xóa Bài Viết thất bại. Lỗi kết nối csdl !')</script>");
+            post = new PostBLL();
+            tagsrelationships = new Tags_relationshipsBLL();
+            post_category_relationships = new Post_Category_relationshipsBLL();
+            int postID = Convert.ToInt32((gwPostmanager.Rows[e.RowIndex].FindControl("lblPostID") as Label).Text);
+            bool delTags = this.tagsrelationships.DeleteWithPostId(postID);
+            bool deletepostCT = this.post_category_relationships.DeleteWithPostId(Convert.ToInt32(postID));
+            bool delPost = this.post.DeleteWithPostID(Convert.ToInt32(postID));
+            if (!delTags || !deletepostCT || !delPost)
+            {
+                Response.Write("<script>alert('Xóa Bài Viết thất bại. Lỗi kết nối csdl !')</script>");
+            }
+            else
+            {
+                Response.Redirect(Request.Url.AbsoluteUri);
+            }
         }
         else
         {
-            Response.Redirect(Request.Url.AbsoluteUri);
+            Response.Write("<script>alert('Bạn không có quyền thực hiện chức năng này !')</script>");
         }
-       
+
     }
     protected void dlCategory_SelectedIndexChanged(object sender, EventArgs e)
     {
